requirements.txt
aiosqlite==0.17.0
alembic==1.8.1
asyncpg==0.27.0
databases==0.6.2
fastapi==0.87.0
python-decouple==3.6
sqlmodel==0.0.8
uvicorn==0.18.3